Board

Description

The Board module is responsible for providing squares, boards, transfering them to a string, as well functions to add and remove shapes.

Types

type Square = Maybe Piece

Square is a synonim to Maybe Piece.

It may adopt a two values:

Nothing for empty square.

Just Piece for piece on square.

 Just (Piece White King)

type Board = Array Coord Square

Board is array of Squares where index of this array is coordinate (Coord).

Common Boards

emptyBoard :: Board

emptyBoard represent a blank board with no Pieces.

In other case emptyBoard is array of empty Squares.

initialBoard :: Board

initialBoard returns a board with standart starting position in chess.

Board's base functions

addPiece :: Coord -> Piece -> Board -> Board

addPiece gets a Piece to set, Coordinates and a Board to set on and returns a new Board with a new setted Piece.

This function may generate error, if you try to add Piece over another Piece.

removePiece :: Coord -> Board -> Board

removePiece gets a Coordinates and a Board to remove piece and returns a new Board without removed Piece.

This function may generate error, if you try to remove Piece from empty Square.

setSquare :: Coord -> Square -> Board -> Board

setSquare gets a Coordinates, Square and a Board to change and returns a new Board with changed Square.

This function didn't generate erorrs and may be used then we don't care that contain prevous Square.

Board's representation functions

displayBoard :: Board -> String

displayBoard gets a Board and returns it's String implementation in cute view.

boardToFen :: Board -> String

This function gets a Board and returns it's String implementation in FEN-notation.

boardFromFen :: String -> Board

boardFromFen gets a fen-string and returns a Board for this record.

This function may generate error if fen is not valid.

Size of Board

xranges :: (Char, Char)

This value contains a bounds of Board by x-axis.

by default it is equal to ('a', 'h')

yranges :: (Int, Int)

This value contains a bounds of Board by y-axis.

by default it is equal to (1, 8)