Coord

Description

Coord module contains basics definitions of Coord type and some operations on it.

Types

type Coord = (Char, Int)

This type implements chess coordinates as tuple of Char and Int.

As example, coordinates of square e4 will be looks like this:

 ('e', 4) 

Representation

coordFromFen :: String -> Coord

This function gets a String, contains coordinates and returns Coord value.

As example, for String d5 it's return (d, 5).

coordToFen :: Coord -> String

This function gets a coordinates and returns it's string representation.

As example, for (g,2) it's return g2.