Evaluation

Description

Evaluation module contains functions and types, that implements position scoring and evaluating.

Types

type Score = Int

Type Score implemets objective value of Piece or Position as example.

Main part

evaluate :: Position -> Score

Function evaluate gets a position and returns it's score