Generator

Description

Generator is more important part of engine, that contains dynamical functions for generating moves and legality checking.

Documentation

moves :: Position -> [Move]

moves function gets a position and returns all pseudo-legal moves for play turn.

isLegalPosition :: Position -> Bool

isLegalPosition function gets position and returns boolean value, that shows that position is legal (no mates, all kings are on board).