Move

Description

Move module contains base type Move and function to convert it in humanity-string.

Types

data Move

Move type contains from, to coordinates and a type of move.

So, this is example of Move value:

 Move NormalMove ('b',3) ('f',3)

Constructors

Move

Fields

kind :: MoveType

from :: Coord

to :: Coord

data MoveType

This type contains kinds of moves.

Constructors

NormalMove
NormalMove implements a non-capture common move.

CaptureMove
CaptureMove implements a common caprute.

PromotionMove PieceType
PromotionMove is Pawn-specific move, that contains a type of piece to promote in.

PawnCapture
PawnCapture is Pawn-specific move, that implements a capture move of Pawn.

PawnDoubleMove
PawnDoubleMove is move that can make Pawn from start position.

EnpassantMove
Enpassant is a special chess move, that also is a capture.

LongCastlingMove
Long Castling (to Queen side) is special King move, that also affects Rook on a-file

ShortCastlingMove
Short Castling (to King side) is special King move, that also affects Rook on h-file

Display-functions

displayMove :: Move -> String

displayMove function gets a move and returns it's string representation.

As example:

 displayMove (Move CaptureMove ('e',2) ('e',6)) == "e2xe6"