Piece

Description

Module Piece contains basic types Piece, PieceColor and PieceType. Also in this module defined some usefull function to representing pieces and transform they.

Types

data Piece

Type Piece unions color and type of piece in complete piece, that has color and type

Constructors

Piece PieceColor PieceType

data PieceColor

PieceColor implements basic values of chess - colors White and Black.

Constructors

Black

White

data PieceType

PieceType implements types of pieces, without setting color to them.

Constructors

Pawn

Knight

Bishop

Rook

Queen

King

Extracting functions

pieceType :: Piece -> PieceType

pieceType gets a Piece and returns it's type.

Example:

>>> pieceType (Piece Black Pawn)
Pawn

pieceColor :: Piece -> PieceColor

pieceColor gets a Piece and returns it's color.

Example:

>>> pieceColor (Piece White Queen)
White

Transformation functions

setType :: PieceType -> Piece -> Piece

This function gets a piece, type and returns a new piece with same color, but new type.

Example:

>>> setType (Queen) (Piece White Pawn)
Piece White Queen

Display

displayPiece :: Piece -> String

This function gets a piece and returns it's string representation for displaing on Board.

pieceToFen :: Piece -> String

pieceToFen function gets a piece and returns it's string representation in FEN-notation. It is just one char.

pieceFromFen :: Char -> Piece

pieceFromFen function gets a character and returns Piece, what represent it in FEN-notation.

Uppercase characters is White pieces. Lowercase characters is Black pieces.

Also, that function may generate error, if character isn't FEN-piece.

pieceTypeToFen :: PieceType -> String

pieceTypeToFen just gets PieceType and returns uppercase character, that implements it in FEN-notation.