Position

Description

Module Position contains type Position (more complex version or Board), and some usefull functions to work with positions.

Types

type TurnColor = PieceColor

TurnColor is just synomim to PieceColor and it implement color of side, what will making move in next turn.

data Castling

Castling type implements castlint availability for all directions and all colors in position.

Constructors

Castling

Fields

whiteLong :: Bool

whiteShort :: Bool

blackLong :: Bool

blackShort :: Bool

data Position

Position type is more complex version of Board. It contains board and additional information - side to turn, castling availability, enpassant square and some information about moves count.

Constructors

Position

Fields

board :: Board

Pieces emplacement.

turn :: TurnColor

Active color.

castling :: Castling

Castling availability.

enpassant :: Maybe Coord

If a pawn has just made a two-square move, this is the position behind the pawn. This is recorded regardless of whether there is a pawn in position to make an en passant capture.

halfmoveClock :: Integer

This is the number of halfmoves since the last pawn advance or capture. This is used to determine if a draw can be claimed under the fifty-move rule.

fullmoveNumber :: Integer

The number of the full move. It starts at 1, and is incremented after Black's move.

Making moves

makeMove :: Move -> Position -> Position

This function gets move and position and returns new position, which is obtained after making move.

Display

displayPosition :: Position -> String

displayPosition gets a position and returns it's String implementation in cute view.

positionFromFen :: String -> Position

positionFromFen gets a fen-string and returns a position for this record.

This function may generate error if fen is not valid.

positionToFen :: Position -> String

This function gets a position and returns it's String implementation in FEN-notation.